Updated views. Juggled migrations.
An earlier attempt to add unique=True to the name fields of all 3 models was preventing django from being able to migrate the file, so I rolled back the database and the migrations to their initial project prompt states.

Got menu_list view and template operational.  menu_list view is down to 2 lines of code!